backend/requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
supabase==2.3.4
openai==1.12.0
python-dotenv==1.0.1
pydantic==2.6.1
networkx==3.2.1
pypdf==4.0.1
unstructured==0.12.3  # Para procesar EUR-Lex con alta calidad
pandas==2.2.0
